Network

+++ UniDB/src/main/java/Storage/StorageManager.java
@@ -3,8 +3,8 @@ package Storage;
 import Engine.Command;
 import Engine.CommandType;
 import Models.Student;
-
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
 public class StorageManager {
@@ -39,17 +39,14 @@ public class StorageManager {
     private final ArrayCollection arrayCollection = new ArrayCollection();
 
     public boolean insertOne(Models.Student student) {
-
-        if (ram.writeRecord(student.toStudentRecord())) {
-            return true;
+        if(ram.exists(student.getId())){
+            return false;
+        }else {
+            if (ram.writeRecord(student.toStudentRecord())) {
+                return true;
+            }
+            return false;
         }
-        return false;
-
-
-//        if (linkedListCollection.insertOne(student)) {
-//            return arrayCollection.insertOne(student);
-//        }
-//        return false;
     }
 
     public boolean deleteOne(Long id) {
@@ -57,19 +54,14 @@ public class StorageManager {
             return true;
         }
         return false;
-//        if (linkedListCollection.deleteOne(id)) {
-//            return arrayCollection.deleteOne(id);
-//        }
-//        return false;
     }
 
     public Models.Student findByID(Long id) {
-        Student record = ram.readRecordByID(id).toStudent();
+        StudentRecord record = ram.readRecordByID(id);
         if (record != null) {
-            return record;
+            return record.toStudent();
         }
         return null;
-//        return linkedListCollection.findByID(id);
     }
 
     public List<Student> findAll() {
@@ -80,6 +72,16 @@ public class StorageManager {
         return students;
     }
 
+    public boolean update(Student student){
+        if (ram.exists(student.getId())){
+            if (ram.writeRecord(student.toStudentRecord())){
+                return true ;
+            }
+            return false;
+        }else{
+            return false;
+        }
+    }
     public List<Command> importDataTransaction(String filePath) {
         List<Command> commands = new java.util.ArrayList<>();
         if (filePath == null || filePath.isEmpty()) {
@@ -90,20 +92,55 @@ public class StorageManager {
             return null;
         }
         for (Models.Student student : students) {
-            if (!insertOne(student)) {
-                return null;
-            } else {
-                commands.add(new Command("db", "s", CommandType.DELETE_ONE, new String[]{
-                        "db",
-                        "s",
-                        "deleteOne",
-                        String.valueOf(student.getId())
-                }));
+            if (findByID(student.id) != null){
+                Models.Student st = findByID(student.getId());
+                if (!update(student)){
+                    return null;
+                }
+                else{
+                    commands.add(new Command("db", "s", CommandType.UPDATE, new String[]{
+                            "db",
+                            "s",
+                            "update",
+                            String.valueOf(st.getId()),
+                            student.getName(),
+                            String.valueOf(st.getGpa())
+                    }));
+                }
+            }else{
+                if (!insertOne(student)){
+                    return null;
+                }else{
+                    commands.add(new Command("db", "s", CommandType.DELETE_ONE, new String[]{
+                            "db",
+                            "s",
+                            "deleteOne",
+                            String.valueOf(student.getId())
+                    }));
+                }
             }
         }
         return commands;
     }
 
+    public boolean importData() {
+        List<Student> students = FileReader.loadFile();
+        if (students.isEmpty()) {
+            return false;
+        }
+        for (Models.Student student : students) {
+            if (findByID(student.id) != null){
+                if (!update(student)){
+                    return false;
+                }
+            }else{
+                if (!insertOne(student)){
+                    return false;
+                }
+            }
+        }
+        return true;
+    }
     public boolean importData(String filePath) {
         if (filePath == null || filePath.isEmpty()) {
             return false;
@@ -114,26 +151,29 @@ public class StorageManager {
             return false;
         }
         for (Models.Student student : students) {
-            if (!insertOne(student)) {
-                return false;
+            if (findByID(student.id) != null){
+                if (!update(student)){
+                    return false;
+                }
+            }else{
+                if (!insertOne(student)){
+                    return false;
+                }
             }
         }
         return true;
     }
 
     public long count() {
-//        return arrayCollection.count();
         return ram.getRecordCount();
     }
 
     public double sumOfField(String fieldName) {
         return ram.sumOfFiled(fieldName);
-//        return arrayCollection.sumOfField(fieldName);
     }
 
     public double averageOfField(String fieldName) {
         return ram.averageOfFiled(fieldName);
-//        return arrayCollection.averageOfField(fieldName);
     }
 
     public List<Student> filterByField(String fieldName, String value) {
@@ -143,9 +183,19 @@ public class StorageManager {
             students.add(record.toStudent());
         }
         return students;
-//        return arrayCollection.filter(fieldName, value);
     }
 
+    public List<Student> filterByField(String fieldName,String start,String end){
+        List<StudentRecord> records = ram.filterByFiled(fieldName,start,end);
+        List<Student> students = new ArrayList<>();
+        for(StudentRecord st : records){
+            students.add(st.ToStudent());
+        }
+        return students;
+    }
+    public boolean deleteAll(){
+        return ram.deleteAll();
+    }
 
 
 }
